Finished up the mobile version of the main page
Styled the "Podcast" and "newsletter" sections of the main-container of the front page in mobile version. I changed the colors of the svg icons for the boxes to dark and changed the svg social media icon colors for the footer to white. Changed some fonts around for better and more readable looks. Finished up the footer. Also changed the background color for the website to a lighter gray, among other changes.