Made background changes to formatting of the hmtl and css code
Removed accidental duplicate I had created in the style.css file. Also, moved the <main> to where it should have been. Added a new grid-template-area for the rows to create three rows in an attempt to fix the page height issue. It didn't work so I might end up getting rid of it. Fixed a couple of bugs on the mobile version of the website as well.